+++ patterns/careers-portfolio.php
@@ -28,8 +28,8 @@
 <!-- /wp:paragraph -->
 
 <!-- wp:buttons -->
-<div class="wp-block-buttons"><!-- wp:button {"backgroundColor":"base","textColor":"contrast","style":{"border":{"color":"var(--wp--preset--color--contrast)","width":"1px","radius":"0px"}}} -->
-<div class="wp-block-button"><a class="wp-block-button__link has-contrast-color has-base-background-color has-text-color has-background has-border-color wp-element-button" href="/shop/" style="border-color:var(--wp--preset--color--contrast);border-width:1px;border-radius:0px">Click here</a></div>
+<div class="wp-block-buttons"><!-- wp:button {"className":"is-style-cta"} -->
+<div class="wp-block-button is-style-cta"><a class="wp-block-button__link wp-element-button" href="/shop/">Click here</a></div>
 <!-- /wp:button --></div>
 <!-- /wp:buttons --></div>
 <!-- /wp:group --></div>